requirements.txt
dnslib
pyyaml
requests
gevent
